% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random Sample from Dirichlet Distribution}
\usage{
rdirichlet(n, a)
}
\arguments{
\item{n}{number of samples}

\item{a}{vector or matrix of shape parameters}
}
\description{
Random generation from the Dirichlet distribution.
}
\examples{
rdirichlet(2, c(1,5,3,8))

}
\seealso{
\code{\link{fit_dirichlet}}
}
