% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary.R
\name{stationary}
\alias{stationary}
\title{Precision of stationary distribution for discrete MCMC variables}
\usage{
stationary(
  z,
  N,
  labels,
  sample = 1000,
  epsilon = "1/M",
  cpu = 1,
  method = "arma",
  digits = 6,
  progress = TRUE,
  summary = TRUE
)
}
\arguments{
\item{z}{MCMC output for the discrete indicator variable with numerical,
character, or factor labels (can also be a \code{\link[coda]{mcmc.list}}
or a matrix with one MCMC chain per column).}

\item{N}{the observed \code{\link{transitions}} matrix (if supplied, \code{z} is ignored).
A quadratic matrix with sampled transition frequencies
(\code{N[i,j]} = number of switches from \code{z[t]=i} to \code{z[t+1]=j}).}

\item{labels}{optional: vector of labels for complete set of models
(e.g., models not sampled in the chain \code{z}). If \code{epsilon=0},
this does not affect inferences due to the improper Dirichlet(0,..,0) prior.}

\item{sample}{number of posterior samples to be drawn for the stationary distribution \eqn{\pi}.}

\item{epsilon}{prior parameter for the rows of the estimated transition matrix \eqn{P}:
\eqn{P[i,]} ~ Dirichlet\eqn{(\epsilon, ..., \epsilon)}.
The default \code{epsilon="1/M"} (with M = number of sampled models) provides
estimates close to the i.i.d. estimates and is numerically stable.
The alternative \code{epsilon=0} minimizes the impact of the prior and
renders non-sampled models irrelevant.
If \code{method="iid"} (ignores dependencies), a Dirichlet prior is assumed on the stationary
distribution \eqn{\pi} instead of the rows of the transition matrix \eqn{P}.}

\item{cpu}{number of CPUs used for parallel sampling.
Will only speed up computations for large numbers of models
(i.e., for large transition matrices).}

\item{method}{how to compute eigenvectors:
\itemize{
  \item \code{"arma"} (default): Uses \code{RcppArmadillo::eig_gen}.
  \item \code{"base"}: Uses \code{base::\link[base]{eigen}}, which might be more stable,
                       but also much slower than \code{"arma"} for small transition matrices.
  \item \code{"eigen"}: Uses package \code{RcppEigen::EigenSolver}
  \item \code{"armas"}: Uses sparse matrices with \code{RcppArmadillo::eigs_gen},
                        which can be faster for very large number of models
                        if \code{epsilon=0} (might be numerically unstable).
  \item \code{"iid"}: Assumes i.i.d. sampling of the model indicator variable \code{z}.
                      This is only implemented as a benchmark, because results cannot
                      be trusted if the samples \code{z} are correlated (which is usually
                      the case for transdimensional MCMC output)
 }}

\item{digits}{number of digits that are used for checking whether the first
eigenvalue is equal to 1 (any difference must be due to low numerical precision)}

\item{progress}{whether to show a progress bar (not functional for \code{cpu>1})}

\item{summary}{whether the output should be summarized.
If \code{FALSE}, posterior samples of the stationary probabilities are returned.}
}
\value{
default: a summary for the posterior distribution of the model
posterior probabilities (i.e., the fixed but unknown stationary distribution of \code{z}).
If \code{summary=FALSE}, posterior samples for \eqn{pi} are returned.
}
\description{
Transdimensional MCMC methods include a discrete model-indicator variable \eqn{z}
with a fixed but unknown stationary distribution with probabilities \eqn{\pi}
(i.e., the model posterior probabiltiies). The function \code{stationary}
draws posterior samples to assess the estimation uncertainty of \eqn{\pi}.
}
\details{
The method draws independent posterior samples of the transition matrix \eqn{P}
for the discrete-valued indicator variable \code{z} (usually, a sequence of sampled models).
For each row of the transition matrix, a Dirichlet\eqn{(\epsilon,...,\epsilon)}
prior is assumed, resulting in a conjugate Dirichlet posterior.
For each sample, the eigenvector with eigenvalue 1 is computed and normalized.
These (independent) posterior samples can be used to assess the estimation
uncertainty in the stationary distribution \eqn{pi} of interest
(e.g., the model posterior probabilities) and to estimate the effective sample size
(see \code{\link{summary.stationary}}).
}
\examples{
# data-generating transition matrix
P <- matrix(c(.1,.5,.4,
              0, .5,.5,
              .9,.1,0), ncol = 3, byrow=TRUE)

# input: sequence of sampled models
z <- rmarkov(500, P)
stationary(z)

# input: transition frequencies
N <- transitions(z)
samples <- stationary(N = N, summary = FALSE)

# summaries:
best_models(samples, k = 3)
summary(samples)
}
\seealso{
\code{\link{best_models}}, \code{\link{summary.stationary}}
}
