% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.stationary}
\alias{summary.stationary}
\title{Summarize posterior probabilities and BF estimates}
\usage{
\method{summary}{stationary}(samples, logBF = FALSE)
}
\arguments{
\item{samples}{posterior samples of the stationary distribution (rows = replications; columns = model probabilities)}

\item{logBF}{whether to summarize log Bayes factors instead of Bayes factors}
}
\value{
a list with estimates for \code{"pp"} = model posterior probabilities, \code{"bf"} = Bayes factors, \code{"neff"} = effective sample size
}
\description{
Summarize posterior probabilities and BF estimates
}
\examples{
n <- 100
pp <- LaplacesDemon::rdirichlet(200, n*c(.7,.2,.1))
class(pp) <- "stationary"
summary(pp)
}

