% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCprecision.R
\docType{package}
\name{MCMCprecision-package}
\alias{MCMCprecision}
\alias{MCMCprecision-package}
\title{MCMCprecision: Precision of discrete parameters in transdimensional MCMC}
\description{
MCMCprecision estimates the precision of the posterior model probabilities in transdimensional Markov chain Monte Carlo methods (e.g., reversible jump MCMC or product-space MCMC). This is useful for applications of transdimensional MCMC such as model selection, mixtures with varying numbers of components, change-point detection, capture-recapture models, phylogenetic trees, variable selection, and for discrete parameters in MCMC output in general.
}
\references{
Heck, D. W., Gronau, Q. F., Overstall, A. M., & Wagenmakers, E.-J. (2017). Quantifying uncertainty in transdimensional Markov chain Monte Carlo using discrete Markov models.
}
\author{
Daniel W. Heck
}
