% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.stationary}
\alias{summary.stationary}
\title{Summarize posterior probabilities and BF estimates}
\usage{
\method{summary}{stationary}(object, logBF = FALSE, ...)
}
\arguments{
\item{object}{posterior samples of the stationary distribution (rows = replications; columns = model probabilities)}

\item{logBF}{whether to summarize log Bayes factors instead of Bayes factors}

\item{...}{ignored}
}
\value{
a list with estimates for \code{"pp"} = model posterior probabilities, \code{"bf"} = Bayes factors, \code{"neff"} = effective sample size
}
\description{
Summarize posterior probabilities and BF estimates
}
\examples{
pp <- matrix(runif(100*3),100)
pp <- pp/rowSums(pp)
class(pp) <- "stationary"
summary(pp)
}
\seealso{
\code{\link{stationary}}
}
